test/cases/ostree-ng.sh: use headless install iso
Rather than pass the .ks to virt-install, fully prepare an ISO for
a headless install up-front.

This (along with reasons for replacing the .ks) will be a common
use-case. In particular when creating installers for headless
bare-metal machines by dd-ing the .iso onto a thumbdrive or similar.

This uses mkksiso from the lorax package.