Add Google Compute Engine BYOS image type (gce-byos)
Add `gce-byos` image type for use with Google Cloud Platform, into
RHEL-8.3 and RHEL-8.4 distro. The image is available only for x86_64
architecture, as the only one supported by GCP. "byos" in the name
stands for "Bring Your Own Subscription", which is consistent with how
Google names such image and it means that users must use
subscription-manager to get RHEL content. The image is loosely based on
the RHEL Guest image provided by Google in their marketplace, because
Google does quite a lot of tweaks to the image, which are currently not
possible with osbuild. Main benefit of the image type is that it uses
proper kernel arguments to work with GCE serial console and availability
of the Google guest tools and SDK on the image.

Google RHUI client is not installed on the GCE image.

dnf-automatic is not installed on the GCE image, since it is not
possible to configure it using osbuild stage.

Add documentation about the GCE image type into `image-types/rhel8/`. It
contains detailed information from the research of Google's RHEL Guest
image as well as description of mechanisms available to import images to
GCP.

Add repository definition with Google Guest tools and use it in the
Cloud API test case.

Modify the Cloud API test case to use the new `gce-byos` image type and
also to use the proper repository with Google Guest tools as part of the
compose request.

Use GPT for gce-byos image also on RHEL-8.3. This makes it consistent
with the image provided by Google and with the image produced for
RHEL-8.4.

Add relevant NEWS entry.

Signed-off-by: Tomas Hozza <thozza@redhat.com>