rhel84: only download parent commit for installer
The parent commit would be added to the sources unconditionally. This
is only necessary for the edge-installer image type.

This doesn't technically change the build behaviour of an existing
distro and image type.  It simply avoids unnecessarily downloading an
ostree commit when only the ref is needed.
It does change the 'sources' section of the manifest however.