tools/image-info: don't convert raw images
After change to image-info [1], which shows the format version for qcow2
images, the `image-format` changed from string to a dictionary. However
the `open_image()` function still compares it with string. This causes
`raw` images to be converted by the script again to `raw` format. This
change fixes the issue, so that `raw` images are not converted, but used
as they are.

[1] https://github.com/osbuild/osbuild-composer/commit/5937b9adca96ea4ffcb0865c75ca3e80c5b9b100

Signed-off-by: Tomas Hozza <thozza@redhat.com>