Weldr: add support to import images to GCP
Add `gcp` upload target support in Weldr API. Allow users to specify the
Storage Bucket to be used and the target Region. Authentication with GCP
must be currently configured in advance in the worker's configuration
file. The GCP Project to use is determined from the GCP credentials file
and can be changed only along with the credentials file in the worker
configuration.

Extend the Weldr API GCP test to also upload the image to GCP and test
it there by creating a VM and connecting to it via SSH.

Add NEWS section describing this change.

Signed-off-by: Tomas Hozza <thozza@redhat.com>