test cases: Introduce a directory for regression test cases
All the integration test cases we currently have make sure that
osbuild-composer works the way we expect it to work. The regression test
cases are supposed to verify the bugs that have been fixed do not
reappear. Introduce a new directory, so that we do not mix these two
categories together.